fix(fleet): deliver initial-prompt to Claude Code inbox for fleet sessions
PR #584 fixed PTY timing issues (#540) by skipping PTY delivery for fleet
claude sessions, intending to rely on "dropbox task.md + Claude inbox".
However, only the dropbox write was implemented — the Claude Code inbox
write was never added. Since Claude Code polls its inbox file (not the
dropbox), the initial prompt was silently dropped.

Changes:
- Move write_to_inbox from CLI inject.rs to kild-core fleet.rs
- Call write_to_inbox from deliver_initial_prompt_for_session when
  skip_pty is true (fleet claude sessions)
- Update inject.rs to use the kild-core version
- Add tests for write_to_inbox in fleet.rs
